# openapi/samal.yaml
openapi: 3.0.3
info:
  title: SAMAL API
  version: 0.1.0
  description: Native messenger for Kyrgyzstan. TLS + session auth. E2EE is phase 2.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/chats/{chatID}/messages:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: chatID
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: Messages newest-first }
    post:
      security: [{ bearerAuth: [] }]
      description: Idempotent on client_id. Optimistic UI on the client.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [client_id, type, payload]
              properties:
                client_id: { type: string }
                type: { type: string, enum: [text, photo, file, voice, location] }
                payload: { type: object }
                reply_to_id: { type: string, format: uuid }
                upload_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "201": { description: Message }
  /v1/receipts:
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [message_ids, status]
              properties:
                message_ids:
                  type: array
                  items: { type: string, format: uuid }
                status: { type: string, enum: [delivered, read] }
      responses:
        "200": { description: ok }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT